refactor: ♻️ (api) 共通して使用されるモジュールを common/service に移動
