+++ apps/api/src/common/service/pubsub/pubsub.module.ts
@@ -1,0 +1,9 @@
+import { Global, Module } from '@nestjs/common';
+import { PubSubService } from './pubsub.service';
+
+@Global()
+@Module({
+  providers: [PubSubService],
+  exports: [PubSubService],
+})
+export class PubSubModule {}
