+++ apps/api/src/common/service/pubsub/pubsub.service.ts
@@ -1,0 +1,13 @@
+import { Injectable, Logger } from '@nestjs/common';
+import { PubSub } from 'graphql-subscriptions';
+
+@Injectable()
+export class PubSubService extends PubSub {
+  private readonly logger = new Logger(PubSubService.name);
+
+  constructor() {
+    super();
+
+    this.logger.debug(`${PubSubService.name} constructed`);
+  }
+}
