+++ apps/api/src/main.ts
@@ -1,7 +1,7 @@
 import { Logger } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app/app.module';
-import { EnvService } from './config/env/env.service';
+import { EnvService } from './common/service/env/env.service';
 
 const bootstrap = async () => {
   const logger = new Logger(bootstrap.name);
